Phase 2: Convert pure utilities to TypeScript
Convert rollDice, createPlayers, showIntro, and showSummary — plus the
tests for createPlayers and showSummary. Types imported from
game/types.ts. showSummary keeps a local structural type since callers
only supply name/money/propertyIds.

Flip the start and test scripts to run via tsx so the mixed .js/.ts
tree keeps working through the rest of the migration.
