--- game/showIntro.js
@@ -1,7 +1,0 @@
-export function showIntro(players) {
-console.log(`Players in the game: ${players.map(player => player.name).join(", ")}.
-=================================
-🎲  MONOPOLY GAME STARTED! 🎲
-=================================
-`);
-}
